+++ src/main/java/com/snackgame/server/messaging/push/service/FCMPushService.java
@@ -1,4 +1,4 @@
-package com.snackgame.server.messaging.notification.service;
+package com.snackgame.server.messaging.push.service;
 
 import java.util.List;
 import java.util.concurrent.Future;
@@ -8,16 +8,16 @@ import org.springframework.stereotype.Service;
 
 import com.google.firebase.messaging.FirebaseMessaging;
 import com.google.firebase.messaging.MulticastMessage;
-import com.snackgame.server.messaging.notification.service.dto.DeviceResponse;
-import com.snackgame.server.messaging.notification.service.dto.NotificationDto;
+import com.snackgame.server.messaging.push.service.dto.DeviceResponse;
+import com.snackgame.server.messaging.push.service.dto.NotificationDto;
 
 @Service
 public class FCMPushService implements PushService {
 
-    private final NotificationService notificationService;
+    private final DeviceService deviceService;
 
-    public FCMPushService(NotificationService notificationService) {
-        this.notificationService = notificationService;
+    public FCMPushService(DeviceService deviceService) {
+        this.deviceService = deviceService;
     }
 
     @Override
@@ -28,7 +28,7 @@ public class FCMPushService implements PushService {
 
     private MulticastMessage makeMessage(String title, String body, Long ownerId) {
 
-        List<DeviceResponse> devicesOf = notificationService.getDevicesOf(ownerId);
+        List<DeviceResponse> devicesOf = deviceService.getDevicesOf(ownerId);
 
         MulticastMessage message = MulticastMessage.builder()
                 .addAllTokens(devicesOf.stream().map(DeviceResponse::getToken).collect(Collectors.toList()))
